[MOD] Add option to skip installing tools in the CI. (#1302)

+++ ci/windows-build.ps1
@@ -17,40 +17,48 @@ Remove-Item "external-${env:PLATFORM}-${env:COMPILER}.zip"
 Move-Item external* binaries
 Write-Output "Installed external dependencies"
 
-Write-Output "Installing OpenSSL"
-Invoke-WebRequest "${env:OPENSSL_URL}" -OutFile "OpenSSL.msi"
-Start-Process msiexec.exe -Wait -ArgumentList '/i OpenSSL.msi /l OpenSSL-install.log /qn'
-Remove-Item OpenSSL.msi
-Remove-Item OpenSSL-install.log
-Write-Output "Installed OpenSSL"
+$DOXYGEN_EXECUTABLE = "C:/Program Files/doxygen/bin/doxygen.exe"
 
-Write-Output "Installing Doxygen"
-New-Item -ItemType directory -Path doxygen | Out-Null
-Set-Location doxygen
-Invoke-WebRequest "${env:DOXYGEN_URL}" -OutFile "doxygen.zip"
-7z x doxygen.zip
-Remove-Item doxygen.zip
-Set-Location "${ROOT_DIR}"
-Write-Output "Installed Doxygen"
+if ("${env:INSTALL_TOOLS}" -eq "YES") {
+    Write-Output "Installing OpenSSL"
+    Invoke-WebRequest "${env:OPENSSL_URL}" -OutFile "OpenSSL.msi"
+    Start-Process msiexec.exe -Wait -ArgumentList '/i OpenSSL.msi /l OpenSSL-install.log /qn'
+    Remove-Item OpenSSL.msi
+    Remove-Item OpenSSL-install.log
+    Write-Output "Installed OpenSSL"
 
-Write-Output "Installing Qt"
-New-Item -ItemType directory -Path "${env:QT_EXTRACT_DIR}" | Out-Null
-Set-Location "${env:QT_EXTRACT_DIR}"
-Invoke-WebRequest "${env:QT_URL}" -OutFile "Qt.7z"
-7z x "Qt.7z"
-Remove-Item Qt.7z
-Set-Location "${ROOT_DIR}"
-Write-Output "Installed Qt"
+    Write-Output "Installing Doxygen"
+    New-Item -ItemType directory -Path doxygen | Out-Null
+    Set-Location doxygen
+    Invoke-WebRequest "${env:DOXYGEN_URL}" -OutFile "doxygen.zip"
+    7z x doxygen.zip
+    Remove-Item doxygen.zip
+    Set-Location "${ROOT_DIR}"
+    Write-Output "Installed Doxygen"
+
+    $DOXYGEN_EXECUTABLE = "${ROOT_DIR}/doxygen/doxygen.exe"
+
+    Write-Output "Installing Qt"
+    New-Item -ItemType directory -Path "${env:QT_EXTRACT_DIR}" | Out-Null
+    Set-Location "${env:QT_EXTRACT_DIR}"
+    Invoke-WebRequest "${env:QT_URL}" -OutFile "Qt.7z"
+    7z x "Qt.7z"
+    Remove-Item Qt.7z
+    Set-Location "${ROOT_DIR}"
+    Write-Output "Installed Qt"
+}
 
 New-Item -ItemType directory -Path build | Out-Null
 
 if ("${env:BUILD_OUTSIDE}" -eq "YES") {
     New-Item -ItemType directory -Path C:\build-outside | Out-Null
     Set-Location C:\build-outside
+} else {
+    Set-Location build
 }
 
 Write-Output "Running cmake"
-cmake -DCMAKE_INSTALL_PREFIX="${ROOT_DIR}/build/install" -DDOXYGEN_EXECUTABLE="${ROOT_DIR}/doxygen/doxygen.exe" -DGENERATE_DOCUMENTATION=ON -DWINDOWS_SERVICE=ON -DCMAKE_CUSTOM_CONFIGURATION_TYPES="${env:CONFIGURATION}" -DCOMPRESS_PDB="${env:COMPRESS_PDB}" -DOPENSSL_ROOT_DIR="${env:OPENSSL_ROOT_DIR}" -DUSE_SYSTEM_OPENSSL=ON -G"${env:GENERATOR}" "${ROOT_DIR}"
+cmake -DCMAKE_INSTALL_PREFIX="${ROOT_DIR}/build/install" -DWIX_VALIDATION="${env:WIX_VALIDATION}" -DDOXYGEN_EXECUTABLE="${DOXYGEN_EXECUTABLE}" -DGENERATE_DOCUMENTATION=ON -DWINDOWS_SERVICE=ON -DCMAKE_CUSTOM_CONFIGURATION_TYPES="${env:CONFIGURATION}" -DCOMPRESS_PDB="${env:COMPRESS_PDB}" -DOPENSSL_ROOT_DIR="${env:OPENSSL_ROOT_DIR}" -DUSE_SYSTEM_OPENSSL=ON -G"${env:GENERATOR}" "${ROOT_DIR}"
 
 Write-Output "Running build"
 cmake --build . --config "${env:CONFIGURATION}"
